requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
numpy>=1.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
